requirements.txt
pandas==2.3.0
openpyxl==3.1.5
uvicorn==0.34.0
uv==0.6.6
litellm==1.72.6
langchain==0.3.25
langchain-core==0.3.65
databricks-langchain==0.5.1
langgraph==0.4.8
python-multipart==0.0.20
python-dotenv==1.1.0
fastapi==0.115.12
jinja2==3.1.6
colorlog==6.9.0